backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai>=1.57.0
httpx==0.27.0
pandas==2.2.3
python-dotenv==1.0.0
pydantic==2.10.6
openpyxl==3.1.2

